Refactor and optimize functions and Value.

+++ lib/src/sql/number.rs
@@ -49,7 +49,7 @@ impl From<i32> for Number {
 
 impl From<i64> for Number {
 	fn from(i: i64) -> Self {
-		Number::Int(i as i64)
+		Number::Int(i)
 	}
 }
